seminar/step1_onnx_model/pyproject.toml
[tool.poetry]
name = "lesson4-step1"
version = "0.1.0"
description = "ONNX model conversion for image captioning"
authors = ["MLOps Course"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.0"
transformers = "^4.35.0"
pillow = "^10.0.0"
onnx = "^1.15.0"
onnxruntime = "^1.16.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
